refactor(local-ci): extract provenance.py (R2-1 sixth cut)
Move 3 provenance helpers (normalize_provenance, provenance_summary,
normalize_result) out of local_ci.py into a sibling provenance.py
module. All three are pure functions, no I/O, no globals — clean leaf
module with zero in-package dependencies.

- local_ci.py: 10,907 → 10,870 lines (-37 net).
- provenance.py: 55 lines.
- Re-exported from local_ci.py so external consumers and tests keep
  the same symbol surface.

Tests: 213/213 local_ci tests pass.

Running R2-1 total: 6 modules extracted (state_paths, normalize,
git_helpers, io_utils, footprint, provenance). local_ci.py shrunk
11,285 → 10,870 lines (-415, -3.7%).
